Email project/requirement.txt
pymilvus == 2.3.6
sentence-transformers
numpy
pyyaml
python-dotenv
tqdm